Alapha Wolfy Ver. Upload
Also known as the Alpha Authority Easter Egg Hunt repl.it project. (AAEEH), it was finalized as a 24/7 bot that runs with a stupid purpose to respond to Scrizeebe with "Nya~!" on command. Uploaded on Alaphaseer Testing as Alapha Wolfy is being converted to Alaphaseer. Roots intended to be shown.